Create FlaskServerControlsFirestore.py
This is an alternative solution proposed by @matthewcook-ku on July 6. The server receives a Text Message from the app. Generates response using ParlAI and then pushes both the message and response to Firebase.

The good thing about this solution is that it minimises read/write calls to Firestore. Thus, helping keep costs at a minimum. 

From an iOS perspective it helps out since it's quite dynamic in how it works.